✓ ext
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/branches/yui@2377 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ support_gimic/sdl2/ext/c86ctl/c86ctlc86box.h
@@ -75,8 +75,8 @@ private:
 	CUsbDev m_usb;					/*!< USB */
 	CGuard m_usbGuard;				/*!< The quard of accessing USB */
 	CGuard m_queGuard;				/*!< The quard of que */
-	size_t m_nQueIndex;				/*!< The position in que */
-	size_t m_nQueCount;				/*!< The count in que */
+	UINT m_nQueIndex;				/*!< The position in que */
+	UINT m_nQueCount;				/*!< The count in que */
 	UINT m_que[0x400];				/*!< que */
 	std::vector<Chip3*> m_chips;	/*!< The list of chips */
